REFACTOR routes and design
Se elimina el componente Home y el Quiz pasa a ser el Home, se cambia el nombre del quiz-api a documentation que es mas correcto
